fix(web): #362 stop top-right control overlap on the diagram
Root cause: two component boundaries placed absolutely-positioned controls
in the same top-right corner with no shared anchor — TopologyCanvas rendered
the Share button at `top-4 right-4`, and the layout added the Composition +
gear cluster at `top-4 right-16`. The 48px magic-number offset collided.

Fix: give the diagram's top-right chrome a single owner.
- Move the Share button out of TopologyCanvas into the layout's control
  cluster, so [Share][Composition][gear] lay out in one flex container at
  `right-4`. The canvas now only owns the top-left status + (renderer's)
  bottom-right zoom.
- The cluster belongs to the CLOSED (diagram) state; hide it when the drawer
  is open so it can't stack on the drawer's own header. The drawer header
  gains a gear (→ Settings) next to its close button, so Settings stays
  reachable from a stage.
